backend/pyproject.toml
[project]
name = "membox-backend"
version = "1.0.0"
description = "MemBox - Multimodal Intelligent Memory System"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.31.0",
    "python-multipart>=0.0.12",
    
    # PowerMem - Intelligent Memory System
    "powermem>=0.1.0",
    
    # OpenAI SDK (for Qwen API)
    "openai>=1.0.0",
    
    # Environment variables
    "python-dotenv>=1.0.1",
    
    # Data validation
    "pydantic>=2.0.0",
    
    # Database connection
    "pymysql>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.uv.sources]
# To install powermem from git
# powermem = { git = "https://github.com/oceanbase/powermem.git" }
